refactor: move FloatEncoding to module
